+++ src/components/AccountSelectTestUtils.ts
@@ -1,0 +1,7 @@
+import { fireEvent } from '@testing-library/vue';
+
+class AccountSelectTestUtils {
+  public static openDropdown(element: HTMLElement): void {
+    fireEvent.mouseDown(element);
+  }
+}