+++ src/components/CurrencyInputTestUtils.ts
@@ -1,0 +1,11 @@
+import { fireEvent } from '@testing-library/vue';
+
+export class CurrencyInputTestUtils {
+  public static async enterValue(
+    element: HTMLElement,
+    value: string
+  ): Promise<void> {
+    await fireEvent.update(element, value);
+    element.dispatchEvent(new Event('change'));
+  }
+}